Add files via upload
Average daily streamflow records are employed in a variety of applications. Daily streamflow
data are used to obtain streamflow statistics such as the 100-year flood and the 7-day 10-year low
flow. Daily streamflow sequences are also commonly employed to calibrate watershed models,
investigate reservoir yield and reliability, explore potential impacts of land use and/or climatic
change on a watershed, and aid in water quantity and water quality simulations. In this
assignment, you are asked to read in a sequence of daily streamflows and calculate the annual
average streamflow